Move humanizeToolTitle into shared format utils
