--- frontend/src/components/chat/tools/antigravity/humanizeToolTitle.ts
@@ -1,5 +1,0 @@
-export const humanizeToolTitle = (title: string): string =>
-  title
-    .replace(/^Running\s+/i, '')
-    .replace(/_/g, ' ')
-    .trim();
